# chore(tracing): improve benchmarks for span creation