--- a/benchmarks/base/requirements.txt
+++ b/benchmarks/base/requirements.txt
@@ -3,4 +3,5 @@
 pyperf
 pyyaml
 attrs
+mock
 httpretty==0.9.7