@mabdinur 2023-04-12
`tracer.trace()` benchmarks already exist